+++ src/module/helpers/_module.mjs
@@ -2,4 +2,3 @@ export * as macros from "./macros.mjs";
 export * as utils from "./utils.mjs";
 export {default as DrawSteelSettingsHandler} from "./settings.mjs";
 export {default as DrawSteelSocketHandler} from "./sockets.mjs";
-export * as rolls from "./rolls.mjs";
