+++ src/module/rolls/_module.mjs
@@ -1,0 +1,6 @@
+// Export order matters
+export {DSRoll} from "./base.mjs";
+export {PowerRoll} from "./power.mjs";
+export {ProjectRoll} from "./project.mjs";
+export {DamageRoll} from "./damage.mjs";
+export {SavingThrowRoll} from "./savingThrow.mjs";
