+++ src/module/rolls/damage.mjs
@@ -1,0 +1,6 @@
+import {DSRoll} from "./base.mjs";
+
+/**
+ * Contains damage-specific info like damage types
+ */
+export class DamageRoll extends DSRoll {}
